Better error messages, removed unneeded casts

+++ src/lib.rs
@@ -15,7 +15,7 @@ pub fn citation() {
 
 pub fn carnivorous_amniota(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(dcal as f32, 2.4839053017);
+    let power = f32::powf(dcal, 2.4839053017);
     let mass = 0.0030393430 * power;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
@@ -32,8 +32,8 @@ pub fn carnivorous_amniota(diameter: f32) {
 
 pub fn carnivorous_amphibia(diameter: f32) {
     let dcal: f32 = diameter;
-    let power1 = f32::powi(dcal as f32, 3);
-    let power2 = f32::powi(dcal as f32, 2);
+    let power1 = f32::powi(dcal, 3);
+    let power2 = f32::powi(dcal, 2);
     let mass = (0.0004064349 * power1) - (0.0041616775 * power2) + (0.0147514015 * dcal) - 0.0122201640;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
@@ -46,7 +46,7 @@ pub fn carnivorous_amphibia(diameter: f32) {
 
 pub fn carnivorous_archosauria(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(dcal as f32, 2.3933363596);
+    let power = f32::powf(dcal, 2.3933363596);
     let mass = 0.0056582325 * power;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
@@ -63,7 +63,7 @@ pub fn carnivorous_archosauria(diameter: f32) {
 
 pub fn carnivorous_felidae(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(dcal as f32, 3.7646426827);
+    let power = f32::powf(dcal, 3.7646426827);
     let mass = 0.0001836115 * power;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
@@ -76,7 +76,7 @@ pub fn carnivorous_felidae(diameter: f32) {
 
 pub fn carnivorous_mammalia(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(dcal as f32, 2.6010376216);
+    let power = f32::powf(dcal, 2.6010376216);
     let mass = 0.0029493394 * power;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
@@ -93,7 +93,7 @@ pub fn carnivorous_mammalia(diameter: f32) {
 
 pub fn carnivorous_squamata(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(dcal as f32, 2.3414629141);
+    let power = f32::powf(dcal, 2.3414629141);
     let mass = 0.0004822862 * power;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
@@ -107,7 +107,7 @@ pub fn carnivorous_squamata(diameter: f32) {
 
 pub fn carnivorous_tetrapoda(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(dcal as f32, 2.6157600696);
+    let power = f32::powf(dcal, 2.6157600696);
     let mass = 0.0017879339 * power;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
@@ -124,7 +124,7 @@ pub fn carnivorous_tetrapoda(diameter: f32) {
 
 pub fn herbivorous_afrotheria(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(dcal as f32, 2.7889164772);
+    let power = f32::powf(dcal, 2.7889164772);
     let mass = 0.0024866807 * power;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
@@ -137,7 +137,7 @@ pub fn herbivorous_afrotheria(diameter: f32) {
 
 pub fn herbivorous_amniota(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(dcal as f32, 2.2123196655);
+    let power = f32::powf(dcal, 2.2123196655);
     let mass = 0.0733160890 * power;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
@@ -154,7 +154,7 @@ pub fn herbivorous_amniota(diameter: f32) {
 
 pub fn herbivorous_artiodactyla(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(dcal as f32, 2.3109191553);
+    let power = f32::powf(dcal, 2.3109191553);
     let mass = 0.4817803510 * power;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
@@ -167,7 +167,7 @@ pub fn herbivorous_artiodactyla(diameter: f32) {
 
 pub fn herbivorous_aves(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(dcal as f32, 2.8100206869);
+    let power = f32::powf(dcal, 2.8100206869);
     let mass = 0.0078389131 * power;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
@@ -180,7 +180,7 @@ pub fn herbivorous_aves(diameter: f32) {
 
 pub fn herbivorous_bovidae(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(dcal as f32, 2.1885588774);
+    let power = f32::powf(dcal, 2.1885588774);
     let mass = 0.6525042837 * power;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
@@ -193,7 +193,7 @@ pub fn herbivorous_bovidae(diameter: f32) {
 
 pub fn herbivorous_cervidae(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(dcal as f32, 3.4595227094);
+    let power = f32::powf(dcal, 3.4595227094);
     let mass = 0.0209279246 * power;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
@@ -206,7 +206,7 @@ pub fn herbivorous_cervidae(diameter: f32) {
 
 pub fn herbivorous_lagomorpha(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(dcal as f32, 2.3225288441);
+    let power = f32::powf(dcal, 2.3225288441);
     let mass = 0.0083097861 * power;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
@@ -219,7 +219,7 @@ pub fn herbivorous_lagomorpha(diameter: f32) {
 
 pub fn herbivorous_mammalia(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(dcal as f32, 2.1815554929);
+    let power = f32::powf(dcal, 2.1815554929);
     let mass = 0.0859205551 * power;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
@@ -236,7 +236,7 @@ pub fn herbivorous_mammalia(diameter: f32) {
 
 pub fn herbivorous_marsupialia(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(dcal as f32, 2.2193247015);
+    let power = f32::powf(dcal, 2.2193247015);
     let mass = 0.0224440978 * power;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
@@ -249,7 +249,7 @@ pub fn herbivorous_marsupialia(diameter: f32) {
 
 pub fn herbivorous_placentalia(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(dcal as f32, 2.2424443852);
+    let power = f32::powf(dcal, 2.2424443852);
     let mass = 0.0879148644 * power;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
@@ -262,7 +262,7 @@ pub fn herbivorous_placentalia(diameter: f32) {
 
 pub fn herbivorous_reptilia(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(dcal as f32, 2.8272692414);
+    let power = f32::powf(dcal, 2.8272692414);
     let mass = 0.0076217107 * power;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
@@ -275,7 +275,7 @@ pub fn herbivorous_reptilia(diameter: f32) {
 
 pub fn herbivorous_rodentia(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(dcal as f32, 2.2974489103);
+    let power = f32::powf(dcal, 2.2974489103);
     let mass = 0.0196247359 * power;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
@@ -288,7 +288,7 @@ pub fn herbivorous_rodentia(diameter: f32) {
 
 pub fn omnivorous_amniota(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(dcal as f32, 2.2146716042);
+    let power = f32::powf(dcal, 2.2146716042);
     let mass = 0.0128321404 * power;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
@@ -305,7 +305,7 @@ pub fn omnivorous_amniota(diameter: f32) {
 
 pub fn omnivorous_artiodactyla(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(dcal as f32, 1.0887395414);
+    let power = f32::powf(dcal, 1.0887395414);
     let mass = 1.2390608405 * power;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
@@ -318,7 +318,7 @@ pub fn omnivorous_artiodactyla(diameter: f32) {
 
 pub fn omnivorous_aves(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(dcal as f32, 3.0157553176);
+    let power = f32::powf(dcal, 3.0157553176);
     let mass = 0.0014342026 * power;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
@@ -331,7 +331,7 @@ pub fn omnivorous_aves(diameter: f32) {
 
 pub fn omnivorous_mammalia(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(dcal as f32, 2.0562969771);
+    let power = f32::powf(dcal, 2.0562969771);
     let mass = 0.0223667325 * power;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
@@ -348,7 +348,7 @@ pub fn omnivorous_mammalia(diameter: f32) {
 
 pub fn omnivorous_rodentia(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(dcal as f32, 1.6780088221);
+    let power = f32::powf(dcal, 1.6780088221);
     let mass = 0.0331020696 * power;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
@@ -361,7 +361,7 @@ pub fn omnivorous_rodentia(diameter: f32) {
 
 pub fn unspecified_canidae(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(dcal as f32, 2.0396557894);
+    let power = f32::powf(dcal, 2.0396557894);
     let mass = 0.0205937247 * power;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
@@ -378,7 +378,7 @@ pub fn unspecified_canidae(diameter: f32) {
 
 pub fn unspecified_carnivora(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(dcal as f32, 2.8590076963);
+    let power = f32::powf(dcal, 2.8590076963);
     let mass = 0.0015578285 * power;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
@@ -395,7 +395,7 @@ pub fn unspecified_carnivora(diameter: f32) {
 
 pub fn unspecified_cricetidae(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(1.8766049284, dcal as f32);
+    let power = f32::powf(1.8766049284, dcal);
     let mass = 0.0180745732 * power;
 
     let dmpw = superscript_number(diameter);
@@ -414,7 +414,7 @@ pub fn unspecified_cricetidae(diameter: f32) {
 
 pub fn unspecified_herpestidae(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(1.0767216732, dcal as f32);
+    let power = f32::powf(1.0767216732, dcal);
     let mass = 0.3503054969 * power;
 
     let dmpw = superscript_number(diameter);
@@ -433,7 +433,7 @@ pub fn unspecified_herpestidae(diameter: f32) {
 
 pub fn unspecified_marsupialia(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(dcal as f32, 2.1922202080);
+    let power = f32::powf(dcal, 2.1922202080);
     let mass = 0.0231733783 * power;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
@@ -450,7 +450,7 @@ pub fn unspecified_marsupialia(diameter: f32) {
 
 pub fn unspecified_mustelidae(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(dcal as f32, 2.5225036146);
+    let power = f32::powf(dcal, 2.5225036146);
     let mass = 0.0038273352 * power;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
@@ -467,7 +467,7 @@ pub fn unspecified_mustelidae(diameter: f32) {
 
 pub fn unspecified_sciuridae(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(1.7069919381, dcal as f32);
+    let power = f32::powf(1.7069919381, dcal);
     let mass = 0.0384600726 * power;
 
     let dmpw = superscript_number(diameter);
@@ -486,7 +486,7 @@ pub fn unspecified_sciuridae(diameter: f32) {
 
 pub fn unspecified_testudines(diameter: f32) {
     let dcal: f32 = diameter;
-    let power = f32::powf(dcal as f32, 2.1345692041);
+    let power = f32::powf(dcal, 2.1345692041);
     let mass = 0.0251303814 * power;
 
     println!("{}{}", colorize(CYAN, "Coprolite diameter: "), colorize(BRIGHT_YELLOW, &format!("{diameter} mm")));
